+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/internal/Delete.java
@@ -15,45 +15,68 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-package org.ballerinalang.nativeimpl.file;
+package org.ballerinalang.nativeimpl.internal;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.bre.bvm.BlockingNativeCallableUnit;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BString;
 import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.nativeimpl.file.utils.Constants;
+import org.ballerinalang.nativeimpl.io.utils.IOUtils;
 import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.ReturnType;
+import org.ballerinalang.util.exceptions.BallerinaException;
 
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Comparator;
 
 /**
- * Creates the file at the path specified in the File struct.
+ * Deletes a given file or a directory.
  *
  * @since 0.970.0-alpha1
  */
 @BallerinaFunction(
-        orgName = "ballerina", packageName = "file",
-        functionName = "Path.getPathValue",
+        orgName = "ballerina", packageName = "internal",
+        functionName = "delete",
         args = {
                 @Argument(name = "path", type = TypeKind.STRUCT, structType = "Path", structPackage = "ballerina.file")
         },
-        returnType = {@ReturnType(type = TypeKind.STRING)},
+        returnType = {
+                @ReturnType(type = TypeKind.BOOLEAN),
+                @ReturnType(type = TypeKind.STRUCT, structType = "IOError", structPackage = "ballerina.file")
+        },
         isPublic = true
 )
-public class GetPathValue extends BlockingNativeCallableUnit {
+public class Delete extends BlockingNativeCallableUnit {
 
-    /**
-     * {@inheritDoc}
-     */
     @Override
     public void execute(Context context) {
         BStruct pathStruct = (BStruct) context.getRefArgument(0);
         Path path = (Path) pathStruct.getNativeData(Constants.PATH_DEFINITION_NAME);
-        String pathValue = path.toString();
-        context.setReturnValues(new BString(pathValue));
+        try {
+            if (!Files.exists(path)) {
+                throw new BallerinaException("failed to delete file: file not found: " + path);
+            }
+            if (!delete(path)) {
+                throw new BallerinaException("failed to delete file: " + path);
+            }
+        } catch (IOException ex) {
+            BStruct errorStruct = IOUtils.createError(context, ex.getMessage());
+            context.setReturnValues(errorStruct);
+        }
+        context.setReturnValues();
     }
+
+    private boolean delete(Path path) throws IOException {
+        Files.walk(path)
+                .sorted(Comparator.reverseOrder())
+                .map(Path::toFile)
+                .forEach(File::delete);
+        return !Files.exists(path);
+    }
+
 }
