+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/internal/GetPathValue.java
@@ -16,63 +16,44 @@
  * under the License.
  */
 
-package org.ballerinalang.nativeimpl.file;
+package org.ballerinalang.nativeimpl.internal;
 
 import org.ballerinalang.bre.Context;
-import org.ballerinalang.bre.bvm.BLangVMErrors;
 import org.ballerinalang.bre.bvm.BlockingNativeCallableUnit;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BBoolean;
+import org.ballerinalang.model.values.BString;
 import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.nativeimpl.file.utils.Constants;
 import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.ReturnType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 
 /**
- * Used to check existence of file.
+ * Creates the file at the path specified in the File struct.
  *
  * @since 0.970.0-alpha1
  */
 @BallerinaFunction(
-        orgName = "ballerina", packageName = "file",
-        functionName = "createFile",
+        orgName = "ballerina", packageName = "internal",
+        functionName = "Path.getPathValue",
         args = {
                 @Argument(name = "path", type = TypeKind.STRUCT, structType = "Path", structPackage = "ballerina.file")
         },
-        returnType = {
-                @ReturnType(type = TypeKind.BOOLEAN),
-                @ReturnType(type = TypeKind.STRUCT, structType = "IOError", structPackage = "ballerina.file")
-        },
+        returnType = {@ReturnType(type = TypeKind.STRING)},
         isPublic = true
 )
-public class CreateFile extends BlockingNativeCallableUnit {
+public class GetPathValue extends BlockingNativeCallableUnit {
 
-    private static final Logger log = LoggerFactory.getLogger(CreateFile.class);
-
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public void execute(Context context) {
         BStruct pathStruct = (BStruct) context.getRefArgument(0);
-        Path filePath = (Path) pathStruct.getNativeData(Constants.PATH_DEFINITION_NAME);
-        Path newFile;
-        try {
-            newFile = Files.createFile(filePath);
-            context.setReturnValues(new BBoolean(Files.exists(newFile)));
-        } catch (IOException | UnsupportedOperationException | SecurityException e) {
-            String msg;
-            if (e instanceof SecurityException) {
-                msg = "Permission denied. Failed to create the file: " + filePath.toString();
-            } else {
-                msg = "Failed to create the file: " + filePath.toString();   
-            }
-            log.error(msg, e);
-            context.setReturnValues(BLangVMErrors.createError(context, msg));
-        }
+        Path path = (Path) pathStruct.getNativeData(Constants.PATH_DEFINITION_NAME);
+        String pathValue = path.toString();
+        context.setReturnValues(new BString(pathValue));
     }
 }
