+++ stdlib/ballerina-builtin/src/main/java/org/ballerinalang/nativeimpl/internal/ToAbsolutePath.java
@@ -15,68 +15,59 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.ballerinalang.nativeimpl.file;
+
+package org.ballerinalang.nativeimpl.internal;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.bre.bvm.BlockingNativeCallableUnit;
+import org.ballerinalang.connector.api.BLangConnectorSPIUtil;
 import org.ballerinalang.model.types.TypeKind;
 import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.nativeimpl.file.utils.Constants;
-import org.ballerinalang.nativeimpl.io.utils.IOUtils;
 import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.ReturnType;
-import org.ballerinalang.util.exceptions.BallerinaException;
 
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Comparator;
 
 /**
- * Deletes a given file or a directory.
+ * Creates the file at the path specified in the File struct.
  *
  * @since 0.970.0-alpha1
  */
 @BallerinaFunction(
-        orgName = "ballerina", packageName = "file",
-        functionName = "delete",
+        orgName = "ballerina", packageName = "internal",
+        functionName = "Path.toAbsolutePath",
         args = {
                 @Argument(name = "path", type = TypeKind.STRUCT, structType = "Path", structPackage = "ballerina.file")
         },
         returnType = {
-                @ReturnType(type = TypeKind.BOOLEAN),
-                @ReturnType(type = TypeKind.STRUCT, structType = "IOError", structPackage = "ballerina.file")
+                @ReturnType(type = TypeKind.STRUCT, structType = "Path", structPackage = "ballerina.file")
         },
         isPublic = true
 )
-public class Delete extends BlockingNativeCallableUnit {
+public class ToAbsolutePath extends BlockingNativeCallableUnit {
 
+    /**
+     * Returns the absolute path of the file.
+     *
+     * @param path the path to the file location.
+     * @return the absolute path reference.
+     */
+    private Path getAbsolutePath(Path path) {
+        return path.toAbsolutePath();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public void execute(Context context) {
         BStruct pathStruct = (BStruct) context.getRefArgument(0);
         Path path = (Path) pathStruct.getNativeData(Constants.PATH_DEFINITION_NAME);
-        try {
-            if (!Files.exists(path)) {
-                throw new BallerinaException("failed to delete file: file not found: " + path);
-            }
-            if (!delete(path)) {
-                throw new BallerinaException("failed to delete file: " + path);
-            }
-        } catch (IOException ex) {
-            BStruct errorStruct = IOUtils.createError(context, ex.getMessage());
-            context.setReturnValues(errorStruct);
-        }
-        context.setReturnValues();
+        BStruct absolutePath = BLangConnectorSPIUtil.createBStruct(context, Constants.FILE_PACKAGE, Constants
+                .PATH_STRUCT);
+        absolutePath.addNativeData(Constants.PATH_DEFINITION_NAME, getAbsolutePath(path));
+        context.setReturnValues(absolutePath);
     }
-
-    private boolean delete(Path path) throws IOException {
-        Files.walk(path)
-                .sorted(Comparator.reverseOrder())
-                .map(Path::toFile)
-                .forEach(File::delete);
-        return !Files.exists(path);
-    }
-
 }
